requirements.txt
Django>=4.0
openai
python-dotenv